extracted Discord Bot Class in sperate py file
